reorganize IAM files
